+++ src/Controller/RecipesController.php
@@ -17,11 +17,13 @@ use App\Repository\CategoryRepository;
 use App\Repository\EndorsementRepository;
 use App\Repository\FavoritesRepository;
 use App\Repository\StepRepository;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class RecipesController extends AbstractController
 {
 
-    #[Route('/recipes', name: 'all_recipes',methods:['GET', 'POST'])]
+    #[Route('/', name: 'all_recipes',methods:['GET', 'POST'])]
     public function allRecipes(Request $request, RecipesRepository $repository, CategoryRepository $cr): Response
     {
         $recipes = $repository->findAll();
@@ -40,19 +42,40 @@ class RecipesController extends AbstractController
     }
 
     #[Route('/recipes/edit/{id}', name: 'edit_recipes',methods:['GET','POST'])]
-    public function editRecipes(Recipes $recipe, Request $request, EntityManagerInterface $em, StepRepository $sr): Response
+    public function editRecipes(Recipes $recipe, Request $request, EntityManagerInterface $em, StepRepository $sr, ValidatorInterface $validator): Response
     {
-        if($request->isMethod('POST'))
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $submittedToken = $request->request->get('_token');
+
+        if($request->isMethod('POST') && $this->isCsrfTokenValid('edit', $submittedToken))
         {
             $recipe->setTitle($request->request->get("_title")); // Attribue le titre depuis la requête
             $recipe->setIntroduction($request->request->get("_introduction")); // Attribue le contenu depuis la requête
             $recipe->setPortions($request->request->get("_portions"));
+
+            $old_image = $recipe->getImage();
             
-            // $recipe = $em->getRepository(Recipes::class)->find($recipeId);
-            // $step = $recipe->setContent("_step",[]);
+            $file = $request->files->get("image_");
+            
+            if ($file)
+            {
+                // ajout nouvelle image
+                $fileName = uniqid().'.'.$file->guessExtension();
+                $destination = $this->getParameter('upload_directory');
+
+                $file->move($destination, $fileName);
+                $recipe->setImage($fileName);
+
+                // suppression ancienne image
+                if($old_image)
+                {
+                    $old_filePath = $this->getParameter('upload_directory'). '/' .$old_image;
+                    unlink($old_filePath);
+                }
+            }
+            
             $steps = $request->request->get("_steps");
             $steps_id = $sr->findBy(['recipe_id'=> $recipe]);
-            
             
             $i=0;
             foreach($steps_id as $value)
@@ -62,6 +85,14 @@ class RecipesController extends AbstractController
             }
 
             $em->persist($recipe, $steps_id); // Prépare l'entité $recipe et les étapes à être sauvegardée dans la base de données
+            
+            $errors = $validator->validate([$recipe,$steps_id]);
+            if(count($errors) > 0)
+            {
+                $this->addFlash('error',"erreur d'informations formulaire");
+                return $this->redirectToRoute('all_recipes');
+            }
+
             $em->flush(); // Sauvegarde réellement les données dans la base de données
             $this->addFlash('success',"La recette a bien été modifié");
             return $this->redirectToRoute('all_recipes');
@@ -72,24 +103,35 @@ class RecipesController extends AbstractController
     
 
     #[Route('/recipes/new', name: 'new_recipe', methods: ['GET', 'POST'])]
-    public function createRecipes(Request $request, EntityManagerInterface $em, CategoryRepository $cr,): Response
+    public function createRecipes(Request $request, EntityManagerInterface $em, CategoryRepository $cr,ValidatorInterface $validator): Response
     {
-        
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $categories = $cr->findAll();
         $recipe = new Recipes();
-        // if($form->isSubmitted() && $form->isValid()){
-        if ($request->isMethod('POST')) { // Si la méthode de la requête est POST (c'est-à-dire que le formulaire a été soumis)
+
+        $submittedToken = $request->request->get('_token');
+
+        if($request->isMethod('POST') && $this->isCsrfTokenValid('new', $submittedToken)) { // Si la méthode de la requête est POST (c'est-à-dire que le formulaire a été soumis)
             
             // On récupère les données soumises dans le formulaire et on les attribue à l'entité
             $recipe->setTitle($request->request->get("_title")); // Attribue le titre depuis la requête
-            $recipe->setImage($request->request->get("_image"));
             $recipe->setIntroduction($request->request->get("_introduction")); // Attribue le contenu depuis la requête
             $recipe->setPortions($request->request->get("_portions"));
             $recipe->setDate(new \DateTime('now'));
+            
+            $file = $request->files->get("image_");
+            
+            if ($file) {
+                $fileName = uniqid().'.'.$file->guessExtension();
+                $destination = $this->getParameter('upload_directory');
 
-            $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+                $file->move($destination, $fileName);
+                $recipe->setImage($fileName);
+            }
+
             $user = $this->getUser();
             $recipe->setUserId($user);
+            
 
             // Récupérer l'ID de la catégorie depuis le formulaire
             $categorieId = $request->request->get('id_categorie',[]);
@@ -103,13 +145,10 @@ class RecipesController extends AbstractController
                 // Attribue l'entité categorie a la variable recipe
                 $recipe->addCategory($categorie);
             }
-        
-            
+            $em->persist($recipe); // Prépare l'entité $recipe à être sauvegardée dans la base de données
+
             //création des étapes
             
-            $em->persist($recipe); // Prépare l'entité $recipe à être sauvegardée dans la base de données
-            $em->flush(); // Sauvegarde réellement les données dans la base de données
-
             $array = $request->request->get('_steps',[]);
 
             $j = 1;
@@ -123,31 +162,44 @@ class RecipesController extends AbstractController
                     $step -> setRank($j);
 
                     $em->persist($step);
-                    $em->flush();
                     $j++;
                 }
             }
 
+            // Validation
+            $errors = $validator->validate([$recipe,$step]);
+            if(count($errors) > 0)
+            {
+                $this->addFlash('error',"erreur d'informations formulaire");
+                return $this->redirectToRoute('all_recipes');
+            }
+
+            $em->flush(); // Sauvegarde réellement les données dans la base de données
+
             $this->addFlash('success',"La recette a bien été crée");
-            return $this->redirectToRoute('new_recipe');
+            return $this->redirectToRoute('recipe_show',['id' => $recipe->getId()]);
         }
         return $this->render('recipes/new.html.twig',['categories' => $categories]);
     }
     
 
     #[Route('/recipes/delete/{id}', name:  'recipe_delete', methods: ['POST','GET'])] // La route '/{id}/delete' permet de supprimer une recette
-    public function delete(Recipes $recipe, EntityManagerInterface $em): Response // La méthode delete() permet de supprimer une recette existante
+    public function delete(Recipes $recipe, EntityManagerInterface $em, Request $request): Response // La méthode delete() permet de supprimer une recette existante
     {
-        $em->remove($recipe); // Supprime une recette de la base de données
-        $em->flush(); // Sauvegarde la suppression dans la base de données
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+        $submittedToken = $request->request->get('_token');
 
+        if($request->isMethod('POST') && $this->isCsrfTokenValid('delete_rec', $submittedToken))
+        {
+            $em->remove($recipe); // Supprime une recette de la base de données
+            $em->flush(); // Sauvegarde la suppression dans la base de données
+        }
         return $this->redirectToRoute('all_recipes'); // Redirige vers la liste des recettes après suppression
     }
 
     #[Route('/recipes/{id}', name: 'recipe_show', methods:'GET')]
     public function show(Recipes $recipe,CommentsRepository $commentaireRepository, FavoritesRepository $fr,EndorsementRepository $er,StepRepository $sr): Response
     {
-
         $recipeID = $recipe->getId(); // Récupérer l'ID de la catégorie depuis le formulaire
         $user = $this->getUser(); // Récupérer informations utilisateur connecté
